--- src/screens/FormElementScreen/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FormElementScreen'
