+++ src/screens/FormScreen/useForm.ts
@@ -1,12 +1,12 @@
-import { useForm } from 'react-hook-form'
+import { useForm as useFormRHF } from 'react-hook-form'
 
 import { zodResolver } from '@hookform/resolvers/zod'
 
 import { defaultValues } from './constants'
-import { formElementSchema } from './formElementSchema'
+import { formElementSchema } from './formSchema'
 
-export const useFormElementForm = () => {
-  const methods = useForm({
+export const useForm = () => {
+  const methods = useFormRHF({
     defaultValues,
     resolver: zodResolver(formElementSchema()),
     mode: 'onChange',
