+++ src/screens/index.ts
@@ -2,6 +2,6 @@ export { default as ButtonScreen } from './ButtonScreen'
 export { default as ColorScreen } from './ColorScreen'
 export { default as ComponentScreen } from './ComponentScreen'
 export { default as DialogScreen } from './DialogScreen'
-export { default as FormElementScreen } from './FormElementScreen'
+export { default as FormScreen } from './FormScreen'
 export { default as LoginScreen } from './LoginScreen'
 export { default as TypographyScreen } from './TypographyScreen'
